Refactor to split graphql into assertions and request files.
