+++ tests/testutils/assertions.go
@@ -2,35 +2,13 @@ package testutils
 
 import (
 	"encoding/json"
-	"fmt"
 	"net/http"
 	"net/http/httptest"
-	"strings"
 	"testing"
 
 	"github.com/99designs/gqlgen/graphql"
 	"github.com/stretchr/testify/assert"
-
-	"gqlgen-starwars/handlers"
 )
-
-func NewGraphQLRequest(t *testing.T, query string) *http.Request {
-	gqlQuery := fmt.Sprintf("{\"query\":\"%v\"}", query)
-
-	req, err := http.NewRequest("POST", "/query", strings.NewReader(gqlQuery))
-	if err != nil {
-		t.Fatal(err.Error())
-	}
-
-	req.Header.Set("Content-Type", "application/json")
-
-	return req
-}
-
-func PerformGraphQLRequest(res http.ResponseWriter, req *http.Request, opts ...handlers.Option) {
-	handler := handlers.NewGraphQlHandler(opts...)
-	handler.ServeHTTP(res, req)
-}
 
 func AssertGraphQLData(t *testing.T, res *httptest.ResponseRecorder, expected string) {
 	t.Helper()
